scripts/requirements.txt
onnxruntime
onnxscript